+++ src/main/java/com/FabrizioBauerDev/BackProtoOlimpiadasSanLuis/Repositories/EtapaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EtapaRepository extends JpaRepository<Etapa, Integer> {
+public interface EtapaRepository extends JpaRepository<Etapa, Long> {
 
 }
