+++ src/main/java/com/FabrizioBauerDev/BackProtoOlimpiadasSanLuis/Repositories/OlimpiadaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OlimpiadaRepository extends JpaRepository<Olimpiada,Integer> {
+public interface OlimpiadaRepository extends JpaRepository<Olimpiada,Long> {
 }
