+++ src/main/java/com/FabrizioBauerDev/BackProtoOlimpiadasSanLuis/Repositories/PruebaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PruebaRepository extends JpaRepository<Prueba,Integer> {
+public interface PruebaRepository extends JpaRepository<Prueba,Long> {
 }
